backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
pgvector==0.3.6
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.7.1
eval_type_backport==0.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20
httpx==0.28.1
cryptography==44.0.0
orjson==3.10.15
pytest==8.3.4
pytest-asyncio==0.25.3
ruff==0.9.6
jieba==0.42.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
sentence-transformers==3.4.1
